Conflicts:
	install/win/installer.nsi
	install/win/installerlang.nsh
	src/Basescape/BaseInfoState.cpp
	src/Basescape/BasescapeState.cpp
	src/Basescape/SellState.h
	src/Basescape/SoldierMemorialState.cpp
	src/Basescape/TransferItemsState.h
	src/Battlescape/Map.cpp
	src/Battlescape/TileEngine.cpp
	src/Engine/Game.cpp
	src/Engine/Game.h
	src/Engine/Music.cpp
	src/Engine/Music.h
	src/Engine/Screen.cpp
	src/Engine/SurfaceSet.h
	src/Geoscape/DogfightState.h
	src/Geoscape/GeoscapeState.cpp
	src/Geoscape/GraphsState.cpp
	src/Interface/TextList.cpp
	src/Interface/TextList.h
	src/Ruleset/MapDataSet.h
	src/Ruleset/RuleAlienMission.cpp
	src/Ruleset/RuleAlienMission.h
	src/Ruleset/RuleRegion.cpp
	src/Ruleset/RuleRegion.h
	src/Ruleset/UfoTrajectory.cpp
	src/Ruleset/UfoTrajectory.h
	src/Savegame/AlienMission.cpp
	src/Savegame/AlienMission.h
	src/Savegame/Base.cpp
	src/Savegame/Base.h
	src/Savegame/SavedGame.cpp
	src/Savegame/SavedGame.h
	src/Savegame/Soldier.cpp
	src/Savegame/TerrorSite.cpp
	src/Savegame/TerrorSite.h
	src/Savegame/Ufo.cpp
	src/Savegame/Ufo.h
	src/Savegame/WeightedOptions.cpp
	src/Savegame/WeightedOptions.h
	src/fmath.h

+++ src/Basescape/SellState.cpp
@@ -383,7 +383,7 @@ void SellState::think()
  * @param selected Selected craft.
  * @return Index of the selected craft.
  */
-int SellState::getCraftIndex(size_t selected) const
+size_t SellState::getCraftIndex(size_t selected) const
 {
 	return selected - _soldiers.size();
 }
@@ -935,7 +935,7 @@ enum SellType SellState::getType(size_t selected) const
  * @param selected Currently selected item.
  * @return Index of the selected item.
  */
-int SellState::getItemIndex(size_t selected) const
+size_t SellState::getItemIndex(size_t selected) const
 {
 	return static_cast<int>(selected)
 		- static_cast<int>(_soldiers.size())
