+++ src/Basescape/SoldierMemorialState.cpp
@@ -117,8 +117,8 @@ SoldierMemorialState::SoldierMemorialState(Game* game)
 	_txtDate->setColor(Palette::blockOffset(13)+10);
 	_txtDate->setText(tr("STR_DATE_UC"));
 
-	int lost = _game->getSavedGame()->getDeadSoldiers()->size();
-	int recruited = lost;
+	size_t lost = _game->getSavedGame()->getDeadSoldiers()->size();
+	size_t recruited = lost;
 	for (std::vector<Base*>::iterator
 			i = _game->getSavedGame()->getBases()->begin();
 			i != _game->getSavedGame()->getBases()->end();
