+++ src/Basescape/TransferItemsState.h
@@ -101,7 +101,7 @@ private:
 	/// Gets type of selected item.
 	enum TransferType getType(size_t selected) const;
 	/// Gets item Index.
-	int getItemIndex(size_t selected) const;
+	size_t getItemIndex(size_t selected) const;
 
 
 	public:
