+++ src/Engine/AdlibMusic.cpp
@@ -177,7 +177,7 @@ void AdlibMusic::player(
 		int i = std::min(delay, len);
 		if (i)
 		{
-			float volume = Game::volumeExponent(Options::musicVolume);
+			float volume = static_cast<float>(Game::volumeExponent(Options::musicVolume));
 			YM3812UpdateOne(
 						opl[0],
 						(INT16*)stream,
