+++ src/Engine/Game.h
@@ -91,7 +91,7 @@ private:
 				int music,
 				int ui);
 		/// Adjusts a linear volume level to an exponential one.
-		static float volumeExponent(int volume);
+		static double volumeExponent(int volume);
 
 		/// Gets the game's display screen.
 		Screen* getScreen() const;
