+++ src/Engine/Music.h
@@ -50,7 +50,7 @@ private:
 		/// Loads music from a chunk of memory.
 		virtual void load(
 				const void* data,
-				size_t size);
+				int size);
 
 		/// Plays the music.
 		virtual void play(int loop = -1) const;
