+++ src/Ruleset/RuleRegion.h
@@ -98,7 +98,7 @@ private:
 	/// Weighted list of the different mission types for this region.
 	WeightedOptions _missionWeights;
 	/// Weight of this region when selecting regions for alien missions.
-	unsigned _regionWeight;
+	size_t _regionWeight;
 	/// All the mission zones in this region.
 	std::vector<MissionZone> _missionZones;
 	/// Do missions in the region defined by this string instead.
@@ -132,7 +132,7 @@ private:
 		std::vector<City*>* getCities();
 
 		/// Gets the weight of this region for mission selection.
-		unsigned getWeight() const;
+		size_t getWeight() const;
 		/// Gets the weighted list of missions for this region.
 		const WeightedOptions& getAvailableMissions() const
 		{
@@ -145,7 +145,7 @@ private:
 		}
 
 		/// Gets a random point inside a mission site.
-		std::pair<double, double> getRandomPoint(unsigned site) const;
+		std::pair<double, double> getRandomPoint(size_t site) const;
 
 		/// Gets the maximum longitude.
 		const std::vector<double>& getLonMax() const
