+++ src/Ruleset/UfoTrajectory.h
@@ -35,11 +35,11 @@ namespace OpenXcom
 struct TrajectoryWaypoint
 {
 	/// The mission zone.
-	unsigned zone;
+	size_t zone;
 	/// The altitude to reach.
-	unsigned altitude;
+	size_t altitude;
 	/// The speed percentage ([0..100])
-	unsigned speed;
+	size_t speed;
 };
 
 
@@ -61,7 +61,7 @@ class UfoTrajectory
 
 private:
 	std::string _id;
-	unsigned _groundTimer;
+	size_t _groundTimer;
 	std::vector<TrajectoryWaypoint> _waypoints;
 
 	public:
@@ -94,20 +94,20 @@ private:
 		 * @param wp The waypoint.
 		 * @return The zone index.
 		 */
-		unsigned getZone(unsigned wp) const
+		size_t getZone(size_t wp) const
 		{
 			return _waypoints[wp].zone;
 		}
 
 		/// Gets the altitude at a waypoint.
-		std::string getAltitude(unsigned wp) const;
+		std::string getAltitude(size_t wp) const;
 
 		/**
 		 * Gets the speed percentage at a waypoint.
 		 * @param wp The waypoint.
 		 * @return The speed as a percentage.
 		 */
-		float getSpeedPercentage(unsigned wp) const
+		float getSpeedPercentage(size_t wp) const
 		{
 			return _waypoints[wp].speed / 100.f;
 		}
@@ -116,7 +116,7 @@ private:
 		 * Gets the number of seconds UFOs should spend on the ground.
 		 * @return The number of seconds.
 		 */
-		unsigned groundTimer() const
+		size_t groundTimer() const
 		{
 			return _groundTimer;
 		}
