+++ src/Savegame/Base.cpp
@@ -2270,15 +2270,15 @@ bool isCompleted::operator()(const BaseFacility* facility) const
 
 
 /**
- * Calculates the detection chance of a base.
+ * Calculates the chance for aLiens to detect this base.
  * Big bases without mindshields are easier to detect.
- * @param difficulty - the current game's difficulty
+ * @param difficulty - the current difficulty setting
  * @return, detection chance
  */
-/*kL unsigned Base::getDetectionChance(int difficulty) const
+/*kL size_t Base::getDetectionChance(int difficulty) const
 {
-	unsigned mindShields = std::count_if(_facilities.begin(), _facilities.end(), isMindShield());
-	unsigned completedFacilities = std::count_if(_facilities.begin(), _facilities.end(), isCompleted());
+	size_t mindShields = std::count_if(_facilities.begin(), _facilities.end(), isMindShield());
+	size_t completedFacilities = std::count_if(_facilities.begin(), _facilities.end(), isCompleted());
 	return ((completedFacilities / 6 + 15) / (mindShields + 1)) * (int)(1 + difficulty / 2);
 } */
 // kL_begin: rewrite getDetectionChance()
