+++ src/Savegame/TerrorSite.h
@@ -41,7 +41,7 @@ class TerrorSite
 private:
 	bool _inBattlescape;
 	int _id;
-	unsigned _secondsRemaining;
+	size_t _secondsRemaining;
 	std::string _race;
 
 
@@ -67,9 +67,9 @@ private:
 		std::wstring getName(Language* lang) const;
 
 		/// Gets the seconds until this terror site expires.
-		unsigned getSecondsRemaining() const;
+		size_t getSecondsRemaining() const;
 		/// Sets the seconds until this terror site expires.
-		void setSecondsRemaining(unsigned seconds);
+		void setSecondsRemaining(size_t seconds);
 
 		/// Gets the terror site's alien race.
 		std::string getAlienRace() const;
