+++ src/Savegame/Ufo.cpp
@@ -161,7 +161,7 @@ void Ufo::load(
 	_direction			= node["direction"].as<std::string>(_direction);
 	_detected			= node["detected"].as<bool>(_detected);
 	_hyperDetected		= node["hyperDetected"].as<bool>(_hyperDetected);
-	_secondsRemaining	= node["secondsRemaining"].as<int>(_secondsRemaining);
+	_secondsRemaining	= node["secondsRemaining"].as<size_t>(_secondsRemaining);
 	_inBattlescape		= node["inBattlescape"].as<bool>(_inBattlescape);
 
 	double
@@ -205,7 +205,7 @@ void Ufo::load(
 
 	std::string tid		= node["trajectory"].as<std::string>();
 	_trajectory			= ruleset.getUfoTrajectory(tid);
-	_trajectoryPoint	= node["trajectoryPoint"].as<unsigned>(_trajectoryPoint);
+	_trajectoryPoint	= node["trajectoryPoint"].as<size_t>(_trajectoryPoint);
 
 	if (_inBattlescape) setSpeed(0);
 }
@@ -380,7 +380,7 @@ void Ufo::setDetected(bool detected)
  * crashed.
  * @return Amount of seconds.
  */
-int Ufo::getSecondsRemaining() const
+size_t Ufo::getSecondsRemaining() const
 {
 	return _secondsRemaining;
 }
@@ -391,7 +391,7 @@ int Ufo::getSecondsRemaining() const
  * crashed.
  * @param seconds Amount of seconds.
  */
-void Ufo::setSecondsRemaining(int seconds)
+void Ufo::setSecondsRemaining(size_t seconds)
 {
 	_secondsRemaining = seconds;
 }
